Rename SingleServerEdgeCursor to DBServerEdgeCursor
