+++ arangod/Graph/Cursors/CMakeLists.txt
@@ -1,2 +1,3 @@
 target_sources(arango_graph PRIVATE
+  DBServerEdgeCursor.cpp
   RefactoredSingleServerEdgeCursor.cpp)
