+++ arangod/Graph/ShortestPathOptions.cpp
@@ -28,7 +28,7 @@
 #include "Cluster/ClusterEdgeCursor.h"
 #include "Cluster/ClusterMethods.h"
 #include "Graph/ClusterTraverserCache.h"
-#include "Graph/SingleServerEdgeCursor.h"
+#include "Graph/Cursors/DBServerEdgeCursor.h"
 #include "Indexes/Index.h"
 #include "Transaction/Helpers.h"
 
@@ -203,7 +203,7 @@ std::unique_ptr<EdgeCursor> ShortestPathOptions::buildCursor(bool backward) {
     return std::make_unique<ClusterShortestPathEdgeCursor>(this, backward);
   }
 
-  return std::make_unique<SingleServerEdgeCursor>(
+  return std::make_unique<DBServerEdgeCursor>(
       this, _tmpVar, nullptr,
       backward ? _reverseLookupInfos : _baseLookupInfos);
 }
